refactor: Add a return value to the job bundle callback
In anticipation of
aws-deadline/deadline-cloud#673,
"fix!: Implement improved job attachments containment warning."
The return value provides the parameter values for use as known paths in
the warning message.

This change is backwards-compatible, because before that PR, the return
value is ignored, and after that PR, the return value will be used to
improve the message.

Updated the reference output parameter_values for all the job bundle
output tests except renderman, which failed to load for me in Maya 2025
+ Renderman 26.2. Removed renderman property references from the other
  test files.

Signed-off-by: Wiebe <399551+mwiebe@users.noreply.github.com>